+++ NEWS.md
@@ -1,7 +1,8 @@
-# lstcompr 0.1.1
+# lstcompr 0.2.0
 
-* variable ranges can also be of character type
-* fixed sorting of free variables
+* variable ranges can also be of character type (fixes https://github.com/patrickroocks/listcompr/issues/1)
+* fixed sorting of free variables (fixes https://github.com/patrickroocks/listcompr/issues/2)
+* auto-detect column names when generating data frames where a symbol is in the base expression
 
 # lstcompr 0.1.0